[Feature] Update Score and ScoreDto to use double for Duration; add ValueValidationFailed method in ErrorUtils

+++ Vereinsmanager.Server.Core/Database/ScoreManagment/Score.cs
@@ -20,7 +20,7 @@ public class Score : MetaData
     [MaxLength(255)]
     public string? Link { get; set; }
 
-    public int? Duration { get; set; }
+    public double? Duration { get; set; }
     
     public ICollection<MusicSheet>? MusicSheets { get; set; }
 }
